requirements.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
streamlit>=1.20.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.2
